requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.5
Django==3.2.7
django-crispy-forms==1.12.0
idna==3.2
lxml==4.6.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
psycopg2==2.9.3
python-dateutil==2.8.2
python-pptx==0.6.20
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.6
XlsxWriter==3.0.1
